requirements.txt
pandas
pendulum
minio
psycopg2-binary
apache-airflow-providers-postgres
cryptography
astronomer-cosmos[dbt.postgres]